feature: new folder structure and code federation
