+++ src/interfaces/product.interface.ts
@@ -1,0 +1,5 @@
+export interface ProductInterface {
+	name: String;
+	image: String;
+	stock: Number;
+}
