+++ src/routes/product.routes.ts
@@ -1,0 +1,9 @@
+import express from 'express';
+import { getProducts, createProduct } from '../controllers/product.controller';
+
+const router = express.Router();
+
+router.get('/', getProducts);
+router.post('/', createProduct);
+
+export const productRouter = router;
